version_1.1.0

+++ src/main/java/ru/hogwarts/school/controller/StudentServiceController.java
@@ -64,7 +64,7 @@ public class StudentServiceController {
     @GetMapping("/get-by-age/{age}")
     @Operation(summary = "Поиск студентов по возрасту",
             description = "Показывает всех студентов, соответсвующих возрасту запроса")
-    public ResponseEntity<List<Student>> getAllStudentsByAge(@PathVariable long age) {
+    public ResponseEntity<List<Student>> getAllStudentsByAge(@PathVariable int age) {
         List<Student> findStudents = service.getStudentsByAge(age);
         if (findStudents == null) {
             return ResponseEntity.notFound().build();
